script1/requirements.txt
openai>=1.0.0
requests>=2.28.1
beautifulsoup4>=4.11.1
trafilatura>=1.5.0
sentence-transformers==4.1.0
faiss-cpu>=1.7.4
rich>=13.3.1
tiktoken>=0.4.0
tenacity>=8.2.2
numpy>=1.22.4
nltk>=3.8.1
lxml_html_clean
python-wordpress-xmlrpc>=2.3
python-dotenv>=1.0.0
google-api-python-client>=2.108.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
google-search-results>=2.4.2
markdown>=3.4.3
tabulate>=0.9.0
tqdm>=4.66.1 
selenium
termcolor
webdriver_manager
g4f
Pillow>=11.0.0  # For image compression
ddgs
huggingface_hub
diffusers